+++ internal/app/diagnose/trace/tracer_test.go
@@ -41,7 +41,7 @@ func testSnapshot() *asset.Snapshot {
 
 func TestPolicyTracer_Trace_Text(t *testing.T) {
 	tracer := &PolicyTracer{}
-	result, err := tracer.Trace(TraceRequest{
+	result, err := tracer.Trace(&TraceRequest{
 		Control:    testControl(),
 		Snapshot:   testSnapshot(),
 		TargetID:   "aws:s3:::test-bucket",
@@ -61,7 +61,7 @@ func TestPolicyTracer_Trace_Text(t *testing.T) {
 
 func TestPolicyTracer_Trace_JSON(t *testing.T) {
 	tracer := &PolicyTracer{}
-	result, err := tracer.Trace(TraceRequest{
+	result, err := tracer.Trace(&TraceRequest{
 		Control:    testControl(),
 		Snapshot:   testSnapshot(),
 		TargetID:   "aws:s3:::test-bucket",
@@ -81,7 +81,7 @@ func TestPolicyTracer_Trace_JSON(t *testing.T) {
 
 func TestPolicyTracer_Trace_ReturnsResult(t *testing.T) {
 	tracer := &PolicyTracer{}
-	result, err := tracer.Trace(TraceRequest{
+	result, err := tracer.Trace(&TraceRequest{
 		Control:    testControl(),
 		Snapshot:   testSnapshot(),
 		TargetID:   "aws:s3:::test-bucket",